==> 05 Model 2020.ipynb <==
# Load packages

# Load data

import numpy as np
import pandas as pd

import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline

from collections import Counter

from sklearn.neural_network import MLPClassifier
from sklearn.preprocessing import StandardScaler
from sklearn.ensemble import RandomForestClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import GridSearchCV, cross_val_score, StratifiedKFold, learning_curve, train_test_split
from sklearn.metrics import brier_score_loss

import xgboost as xgb
from xgboost import XGBClassifier

#all_games_df = pd.read_csv('Data/~Created data/all_games_df.csv')

path = "C:/Users/anovitt/AppData/Local/Programs/Python/Python36/Scripts/kaggle_mania_2020_men/google-cloud-ncaa-march-madness-2020-division-1-mens-tournament/"

#data_matrix_df = pd.read_csv('data_matrix.csv')

data_matrix_df = pd.read_csv(path + 'data_matrix.csv')
sub_matrix_df = pd.read_csv(path + 'sub_matrix.csv')

data_matrix_df.head(3)

XX_train = data_matrix_df.drop(['Season','DayNum','T1','T1_Points','T2','T2_Points','Location','ResultDiff','T1_AP','T2_AP'],axis =1)

XX_train.head(3)

sub_matrix_df.head(3)

X_test = sub_matrix_df.drop(['Season','ID','T1','T2','Pred','Unnamed: 0','T1_AP','T2_AP'],axis =1)

X_test.head(3)


data_matrix_df['Win'] = 0

data_matrix_df.loc[data_matrix_df['T1_Points'] > data_matrix_df['T2_Points']] = 1


yy_train = data_matrix_df['Win']

X_train, X_val, y_train, y_val = train_test_split(XX_train, yy_train, test_size=0.25, random_state=42)


#X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.33, random_state=42)

Note: test is 2019 predictions but our "test" holdout set is referred to as "valid" 

X_test = X_test.astype("float64")
XX_train = XX_train.astype("float64") # Full data set, use for round 2
X_train = X_train.astype("float64")
X_val = X_val.astype("float64")
#y_train = y_train.astype("float64")
#y_val = y_val.astype("float64")

# Scoring rules and two benchmarks

def LogLoss(predictions, realizations):
    predictions_use = predictions.clip(0)
    realizations_use = realizations.clip(0)
    LogLoss = -np.mean( (realizations_use * np.log(predictions_use)) + 
                        (1 - realizations_use) * np.log(1 - predictions_use) )
    return LogLoss

# Fit a neural network (with and without line)

Normalize data (using z-scores) before neural network

scaler = StandardScaler()
scaler.fit(X_train)  # Fit only to the training data
scaler.fit(X_val)
scaler.fit(X_test)
scaler.fit(XX_train)
scaled_X_train = pd.DataFrame(scaler.transform(X_train), index=X_train.index, columns=X_train.columns)
scaled_XX_train = pd.DataFrame(scaler.transform(XX_train), index=XX_train.index, columns=X_train.columns) # full data set
scaled_X_valid = pd.DataFrame(scaler.transform(X_val), index=X_val.index, columns=X_val.columns)
scaled_X_test = pd.DataFrame(scaler.transform(X_test), index=X_test.index, columns=X_test.columns)

With line (note: we won't have line in the rounds after the first, but we could use this for the first round only like Lopez and Matthews did) 

#Note: I tried logistic activation and different combinations of hidden layers/nodes
#Hyperparameters below minimized the log loss in the holdout set
#I also submited a prediction with 10 nodes in the first layer, but this is the submission that placed 4th (w/ 8 in 1st)
#nn = MLPClassifier(activation='relu', hidden_layer_sizes=(8,5,3),random_state=201, max_iter=1000)
nn = MLPClassifier(activation='relu', hidden_layer_sizes=(100,50,25,5,3),random_state=201, max_iter=1000)
nn.fit(scaled_X_train,y_train)
nn_pred = pd.DataFrame(nn.predict_proba(scaled_X_valid))[1]

LogLoss(nn_pred, y_val)

# Full data set, use for round 2
nn = MLPClassifier(activation='relu', hidden_layer_sizes=(8,8,8,5,3),random_state=201, max_iter=1000)
nn.fit(scaled_XX_train,yy_train)
nn_pred = pd.DataFrame(nn.predict_proba(scaled_X_test))[1]

nn.fit(scaled_XX_train,yy_train)
nn_pred = pd.DataFrame(nn.predict_proba(scaled_X_test))[1]

LogLoss(nn_pred, y_val)

#try second holdout (does worse, but still better than baseline of 54)
#nn = MLPClassifier(activation='relu', hidden_layer_sizes=(8,8,8,5,3),random_state=201, max_iter=1000)
#nn.fit(scaled_XX_train,yy_train)
nn_pred_test = pd.DataFrame(nn.predict_proba(scaled_X_test))[1]

# write out prob csv file

nn_pred.to_csv(path + 'submitProb.csv', index=False)

***1D CNN Netwok Try***

# CNN 1D Network

#import keras

from tensorflow.keras import backend
#import keras

#from matplotlib import pyplot
#from keras.models import Sequential
#from keras.layers import Dense
#from keras.layers import Flatten
#from keras.layers import Dropout
#from keras.layers.convolutional import Conv1D
#from keras.layers.convolutional import MaxPooling1D
#from keras.utils import to_categorical

scaled_X_train.shape

def evaluate_model(X_train, y_train, X_val, y_val):
    
    verbose, epochs, batch_size = 0, 10, 32
    n_timesteps = 1 #X_train.shape[2]
    n_features = X_train.shape[2]
    n_outputs = 1 #y_train.shape[1]
    
    #n_timesteps, n_features, n_outputs = X_train.shape[1], X_train.shape[2], Y_train.shape[1]
    model = Sequential()
    model.add(Conv1D(filters=64, kernel_size=3, activation='relu', input_shape=(n_timesteps,n_features)))
    model.add(Conv1D(filters=64, kernel_size=3, activation='relu'))
    model.add(Dropout(0.5))
    model.add(MaxPooling1D(pool_size=2))
    model.add(Flatten())
    model.add(Dense(100, activation='relu'))
    model.add(Dense(n_outputs, activation='softmax'))
    model.compile(loss='categorical_crossentropy', optimizer='adam', metrics=['accuracy'])
    # fit network
    model.fit(X_train, y_train, epochs=epochs, batch_size=batch_size, verbose=verbose)
    # evaluate model
    _, accuracy = model.evaluate(X_val, y_val, batch_size=batch_size, verbose=0)
    return accuracy

# run an experiment
def run_experiment(repeats=10):
    # load data
	
	# repeat experiment
	scores = list()
	for r in range(repeats):
		score = evaluate_model(scaled_X_train,y_train, scaled_X_valid, y_val)
		score = score * 100.0
		print('>#%d: %.3f' % (r+1, score))
		scores.append(score)
	# summarize results
	summarize_results(scores)

run_experiment()

print('hello')

Without line

# Create test predictions

#Note: I'm predicting home (lower seed) win probability. Need to convert to be consistent with output file (lower team ID)
second_rd_submission = pd.merge(test_combos_df, second_rd_submission_all, left_index=True, right_index=True)

second_rd_submission.loc[second_rd_submission['HTeamID']<second_rd_submission['RTeamID'], 'pred'] = second_rd_submission[1]
second_rd_submission.loc[second_rd_submission['HTeamID']>second_rd_submission['RTeamID'], 'pred'] = second_rd_submission[0]

second_rd_submission.to_csv('Data/~Created data/second_rd_submission_all.csv', index=False)

second_rd_submission = second_rd_submission[['ID','pred']]

second_rd_submission.head()

second_rd_submission.tail()

#Export to submit to Kaggle
second_rd_submission.to_csv('Data/~Created data/second_rd_submission.csv', index=False)

# Other models

No model performed as well as the neural network

XGBoost
